Change IFormatter to use Number instead of Double by default.

+++ cayenne/src/test/java/nl/sikken/bertrik/cayenne/SimpleCayenneTest.java
@@ -40,9 +40,9 @@ public final class SimpleCayenneTest {
         // decode it
         CayenneMessage message = CayenneMessage.parse(data);
         Assert.assertEquals(12, message.getItems().size());
-        Assert.assertEquals(52.0, message.ofType(ECayenneItem.GPS_LOCATION).getValues()[0], 0.1);
-        Assert.assertEquals(3.82, message.ofType(ECayenneItem.ANALOG_INPUT).getValue(), 0.01);
-        Assert.assertEquals(19.0, message.ofType(ECayenneItem.TEMPERATURE).getValue(), 0.1);
+        Assert.assertEquals(52.0, message.ofType(ECayenneItem.GPS_LOCATION).getValues()[0].doubleValue(), 0.1);
+        Assert.assertEquals(3.82, message.ofType(ECayenneItem.ANALOG_INPUT).getValue().doubleValue(), 0.01);
+        Assert.assertEquals(19.0, message.ofType(ECayenneItem.TEMPERATURE).getValue().doubleValue(), 0.1);
     }
     
     /**
